+++ libs/providers/langchain-baseten/src/chat_models.ts
@@ -12,12 +12,15 @@ import { ChatOpenAI } from "@langchain/openai";
 import type { LangSmithParams } from "@langchain/core/language_models/chat_models";
 import type { BaseMessage } from "@langchain/core/messages";
 import { AIMessageChunk } from "@langchain/core/messages";
-import type { ToolCallChunk } from "@langchain/core/messages/tool";
 import type { ChatGenerationChunk, ChatResult } from "@langchain/core/outputs";
 import type { CallbackManagerForLLMRun } from "@langchain/core/callbacks/manager";
 import { getEnvironmentVariable } from "@langchain/core/utils/env";
 import { DEFAULT_BASE_URL, DEFAULT_API_KEY_ENV_VAR } from "./const.js";
 import { normalizeModelUrl } from "./converters/params.js";
+import {
+  normalizeToolCallChunks,
+  chunkHasContent,
+} from "./converters/chat_models.js";
 import type { BasetenChatInput } from "./types.js";
 
 function resolveApiKey(fields?: BasetenChatInput): string {
@@ -33,56 +36,6 @@ function resolveApiKey(fields?: BasetenChatInput): string {
       `option or set the ${DEFAULT_API_KEY_ENV_VAR} environment variable.\n\n` +
       `  Get your API key at https://app.baseten.co/settings/api-keys`
   );
-}
-
-/**
- * Fix TensorRT-LLM tool-call streaming quirks:
- * - Fold same-index deltas within a single SSE event into one entry
- * - Clear `id` on continuation deltas (no `name`) so `concat()` merges by index
- *
- * See: Python `langchain-baseten._normalize_tool_call_chunks`
- */
-export function normalizeToolCallChunks(
-  chunks: ToolCallChunk[]
-): ToolCallChunk[] {
-  if (chunks.length <= 1 && (!chunks[0] || chunks[0].name)) return chunks;
-
-  const byIndex = new Map<number, ToolCallChunk>();
-
-  for (const tc of chunks) {
-    if (tc.index == null) continue;
-    const existing = byIndex.get(tc.index);
-    if (!existing) {
-      byIndex.set(tc.index, { ...tc });
-    } else {
-      byIndex.set(tc.index, {
-        ...existing,
-        name: existing.name ?? tc.name,
-        args: (existing.args ?? "") + (tc.args ?? ""),
-        id: existing.id ?? tc.id,
-      });
-    }
-  }
-
-  const result: ToolCallChunk[] = [];
-  for (const tc of byIndex.values()) {
-    if (!tc.name && tc.id != null) {
-      result.push({ ...tc, id: undefined });
-    } else {
-      result.push(tc);
-    }
-  }
-
-  return result;
-}
-
-function chunkHasContent(message: AIMessageChunk): boolean {
-  if (typeof message.content === "string" && message.content.length > 0)
-    return true;
-  if (Array.isArray(message.content) && message.content.length > 0) return true;
-  if (message.tool_call_chunks && message.tool_call_chunks.length > 0)
-    return true;
-  return false;
 }
 
 function inferModelNameFromUrl(url: string): string {
