+++ libs/providers/langchain-baseten/src/index.ts
@@ -1,3 +1,4 @@
-export { ChatBaseten, normalizeToolCallChunks } from "./chat_models.js";
+export { ChatBaseten } from "./chat_models.js";
+export { normalizeToolCallChunks } from "./converters/chat_models.js";
 export { normalizeModelUrl } from "./converters/params.js";
 export type { BasetenChatInput } from "./types.js";
